Moved missing to love.d.ts
