+++ index.d.ts
@@ -1,6 +1,5 @@
-import "./love";
 import "./debug";
 import "./global";
-import "./missing";
+import "./love";
 import "./math";
 import "./os";
